Add source PDFs and evaluation results
- Created README.md in the data/raw directory to document the source PDFs used for ingestion, including naming conventions and examples of files.
- Added examples_qualitative.csv to the evaluation directory containing qualitative evaluation prompts and scores.
- Introduced results_A_embedders.csv with performance metrics for different embedding models.
- Added results_B_llms.csv with performance metrics for various language models.
- Included results_C_prompts.csv summarizing keyword performance for different prompt types.

+++ src/generation.py
@@ -26,7 +26,7 @@ class LLMGenerator:
             )
             kwargs["device_map"] = "auto"
         elif self.device == "mps":
-            kwargs["torch_dtype"] = torch.float16
+            kwargs["torch_dtype"] = torch.bfloat16
         else:
             kwargs["torch_dtype"] = torch.float32
